I did the overall launch, register and login page. The Frontend.py is used to run the code.
I need functions validate_login(username, password) which needs to be implemented by backend. It should check if username exists, then check if password is correct. This function should return a False if either or both requirements are not fulfilled. Else it should return true. I also need a validate_register(username, password) which returns False if username and password does not match the requirements and returns True if it fulfils the requirements